lsf: add a helper to get a list of all jobs
Using the API for retrieving a list of jobs is complicated because
lsb_readjobinfo() takes a reference to an int when retrieving jobs. This patch
fixes this by adding a helper that can be called and that will return a list of
all jobs present in the batch daemon. With this simple interface, users of the
API can, then, filter jobs however they like.

Signed-off-by: Renato L. F. Cunha <renatoc@br.ibm.com>